backend/requirements.txt
# Dépendances principales
fastapi>=0.95.0
uvicorn>=0.22.0
sqlalchemy>=2.0.0
pydantic>=1.10.0
alembic>=1.10.0
psycopg2-binary>=2.9.5
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
email-validator>=2.0.0
python-dotenv>=1.0.0
bcrypt>=4.0.0

# Gestion des API et communication
httpx>=0.24.0
requests>=2.28.0
websockets>=11.0.0

# Outils pour LLM et IA
openai>=0.27.0
langchain>=0.0.200
qdrant-client>=1.0.0
tiktoken>=0.4.0

# Utilitaires
pyyaml>=6.0
jinja2>=3.1.2
tenacity>=8.2.0
pytz>=2023.3

# Monitoring et logging
loguru>=0.7.0
prometheus-client>=0.16.0

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0
